+++ SignNow.Net.Test/FeatureTests/RoleBasedInviteTest.cs
@@ -2,8 +2,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SignNow.Net.Model;
-using SignNow.Net.Test;
 using SignNow.Net.Test.FakeModels;
+using UnitTests;
 
 namespace FeatureTests
 {