+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## Changed
 - Added support for Basic Auth param to `Model.Token`
 
+## Changed
+- Increased timeout for Http Client
+- Added Http method, Url and response time to error messages
+
+
 ## [0.5.2-beta] - 2020-05-22
 ### Fixed
 - Fixed `Models.FieldContents.RadiobuttonContent` converting error [#104](https://github.com/signnow/SignNow.NET/issues/104)